Rollout UI and create a GreedyService
