+++ greedy-timetabling/src/main/resources/META-INF/resources/index.html
@@ -36,11 +36,14 @@
     <p>Generate the optimal schedule for your teachers and students.</p>
 
     <div class="mb-2">
+        <button id="refreshButton" type="button" class="btn btn-secondary">
+            <span class="fas fa-refresh"></span> Refresh
+        </button>
         <button id="solveButton" type="button" class="btn btn-success">
             <span class="fas fa-play"></span> Solve
         </button>
-        <button id="clean" type="button" class="btn btn-danger">
-            <span class="fas fa-stop"></span> Clean solution
+        <button id="stopSolvingButton" type="button" class="btn btn-danger">
+            <span class="fas fa-stop"></span> Stop solving
         </button>
         <span id="score" class="score ml-2 align-middle font-weight-bold">Score: ?</span>
